+++ packages/Webkul/Shop/src/Resources/views/checkout/cart/index.blade.php
@@ -131,7 +131,7 @@
 
                                 @if (! cart()->hasError())
                                     @php
-                                        $minimumOrderAmount = (int) core()->getConfigData('sales.orderSettings.minimum-order.minimum_order_amount') ?? 0;
+                                        $minimumOrderAmount = core()->getConfigData('sales.orderSettings.minimum-order.minimum_order_amount') ?? 0;
                                     @endphp
 
                                     <proceed-to-checkout
